Create Main code
This is supposed to print n number of part lengths the number of times that they are required (don't ask!) but so far I can only get it to do that for the last part length. 
I had it working fine when n was preset but I don't want to have to go through entering 0 as part lengths if I couldn't fill the preset number. I.e when I just had PN = Array.new(3) and used for I in 1..3.